+++ src/expression/boolean/BinaryBooleanOpExpr.ts
@@ -6,7 +6,6 @@ import { IfcExpressionContext } from "../../context/IfcExpressionContext.js";
 import { ExprEvalError } from "../ExprEvalResult.js";
 import { Type, Types } from "../../type/Types.js";
 import { ExprType } from "../../type/ExprType.js";
-import { ExprStringBuilder } from "../ExprStringBuilder.js";
 import { LogicalValue } from "../../value/LogicalValue";
 
 export abstract class BinaryBooleanOpExpr extends Expr2<